Added economy model file for treasury yields
